pyproject.toml
[project]
name = "teufa"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
    {name = "Brian Rosner", email = "brian@brosner.com"},
]
license = {text = "MIT"}
requires-python = "<4.0,>=3.13"
dependencies = [
    "click<9.0.0,>=8.1.7",
    "gunicorn<23.0.0,>=22.0.0",
    "flask<4.0.0,>=3.0.3",
    "flask-sqlalchemy<4.0.0,>=3.1.1",
    "psycopg<4.0.0,>=3.1.19",
    "alembic<2.0.0,>=1.13.1",
    "pydantic<3.0.0,>=2.7.1",
]

[project.scripts]
teufa = "teufa.cli:cli"

[dependency-groups]
test = [
    "pytest<9.0.0,>=8.2.0",
    "pytest-cov<6.0.0,>=5.0.0",
]
dev = [
    "pydeps<2.0.0,>=1.12.20",
    "ruff<1.0.0,>=0.5.0",
]

[tool.uv]
package = true
default-groups = []

[tool.setuptools]
packages = ["teufa"]
